requirements.txt
python-dotenv==0.21.1
aiogram==2.24
openai==0.26.4
chatgpt==2.2212.0
gino==1.0.1
SQLAlchemy==1.3.24
asyncpg==0.27.0
APScheduler==3.10.0